% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{AddClonesToPlot}
\alias{AddClonesToPlot}
\title{AddClonesToPlot}
\usage{
AddClonesToPlot(seuratObj, plot, colorShapes = "black")
}
\arguments{
\item{seuratObj}{The seurat object}

\item{colorShapes}{A string or vector that is passed to geom_point()}

\item{The}{plot object, such as the result from FeaturePlot()}
}
\description{
Can be used to highlight a set of cells from a seurat plot, such as overlaying specific clonotypes
}
