% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_Fixes.R
\name{AddModuleScoreAvg}
\alias{AddModuleScoreAvg}
\title{AddModuleScoreAvg}
\usage{
AddModuleScoreAvg(SeurObj, genes.list = NULL, genes.pool = NULL,
  n.bin = 25, seed.use = 1, ctrl.size = 100,
  enrich.name = "Cluster", random.seed = 1, returnScore = F)
}
\arguments{
\item{genes.list, }{Gene list to obtain a score for}

\item{genes.pool, }{Gene list to base as the pool; NULL = all.}

\item{n.bin, }{number of bins to evaluate score across; default 25.}

\item{seed.use, }{seed use.}

\item{ctrl.size, }{control gene set size.}

\item{enrich.name, }{A name for the assesment}

\item{random.seed, }{random seed}

\item{object, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
Instead of the status quo score of Seurat which is 1 score 1 gene, this function takes a list of genes and computes per set, the average of the individual scores.
}
\keyword{SerIII_AddModuleScore}
