% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{AddTitleToMultiPlot}
\alias{AddTitleToMultiPlot}
\title{AddTitleToMultiPlot}
\usage{
AddTitleToMultiPlot(plotGrid, title, relHeights = c(0.1, 1))
}
\arguments{
\item{plotGrid}{The list of plots}

\item{title}{The title for this plot}

\item{relHeights}{The relative heights, passed to plot_grid}
}
\description{
Add overall title to a plot_grid, as as the results generated by FeaturePlot() from multiple genes
}
