% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAliasing.R
\name{aliasDAVID}
\alias{aliasDAVID}
\title{aliasDAVID}
\usage{
aliasDAVID(
  ensemblIds = NULL,
  geneSymbols = NULL,
  email,
  replaceUnmatched = TRUE
)
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs, passed to the getBM() filters argument}

\item{geneSymbols}{A vector of gene symbols, passed to the getBM() filters argument}

\item{email}{email account registered with DAVIDWebService}

\item{replaceUnmatched}{Logical. If TRUE, removes NA's and replaces with inputs}
}
\description{
aliasDAVID
}
