% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAliasing.R
\name{aliasENSEMBL}
\alias{aliasENSEMBL}
\title{aliasENSEMBL}
\usage{
aliasENSEMBL(
  ensemblIds = NULL,
  geneSymbols = NULL,
  attributes = c("hgnc_symbol", "ensembl_gene_id", "external_gene_name"),
  dataset = "mmulatta_gene_ensembl",
  version = NULL,
  mirror = "uswest",
  biomart = "ensembl",
  replaceUnmatched = TRUE
)
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs, passed to the getBM() filters argument}

\item{geneSymbols}{A vector of gene symbols, passed to the getBM() filters argument}

\item{attributes}{A vector of ensembl attributes}

\item{dataset}{Passed directly to biomaRt::useEnsembl}

\item{version}{Passed directly to biomaRt::useEnsembl}

\item{mirror}{Passed directly to biomaRt::useEnsembl}

\item{biomart}{Passed directly to biomaRt::useEnsembl}

\item{replaceUnmatched}{Logical. If TRUE, removes NA's and replaces with inputs}
}
\description{
aliasENSEMBL
}
