% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNames.R
\name{AliasGeneNames}
\alias{AliasGeneNames}
\title{AliasGeneNames}
\usage{
AliasGeneNames(
  ensemblIds,
  biomart = "ensembl",
  dataset = "mmulatta_gene_ensembl",
  ensemblFilters = c("ensembl_gene_id"),
  ensemblVersion = NULL
)
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs, passed to the getBM() filters argument}

\item{biomart}{Passed directly to biomaRt::useEnsembl}

\item{dataset}{Passed directly to biomaRt::useEnsembl}

\item{ensemblFilters}{A vector of ensembl IDs, passed to the getBM() filters argument.}

\item{ensemblVersion}{Passed directly as version to biomaRt::useEnsembl}
}
\description{
This function accepts a vector of gene names. It will first query Ensembl to attempt to translate ID to gene symbol. Any names not translated will be compared for human homologs. Finally, these symbols will be compared to CD genes, and the CD name appended (if found).
}
