% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAliasing.R
\name{aliasSTRINGdb}
\alias{aliasSTRINGdb}
\title{aliasSTRINGdb}
\usage{
aliasSTRINGdb(
  ensemblIds = NULL,
  geneSymbols = NULL,
  speciesId = 9606,
  replaceUnmatched = TRUE
)
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs, passed to the getBM() filters argument}

\item{geneSymbols}{A vector of gene symbols, passed to the getBM() filters argument}

\item{speciesId}{Species ID. see Stringdb reference for list of avialable species}

\item{replaceUnmatched}{Logical. If TRUE, removes NA's and replaces with inputs}
}
\description{
aliasSTRINGdb
}
