% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAliasing.R
\name{aliasTable}
\alias{aliasTable}
\title{aliasTable}
\usage{
aliasTable(
  ensemblIds = NULL,
  geneSymbols = NULL,
  davidEmail,
  ensemblAttributes = c("hgnc_symbol", "ensembl_gene_id", "external_gene_name"),
  ensemblDataset = "mmulatta_gene_ensembl",
  ensemblVersion = NULL,
  ensemblMirror = "uswest",
  biomart = "ensembl",
  stringSpeciesId = 9606,
  aliasPriorityOrder = c("ENSEMBL", "STRING", "DAVID")
)
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs, passed to the getBM() filters argument}

\item{geneSymbols}{A vector of gene symbols, passed to the getBM() filters argument}

\item{davidEmail}{email account registered with DAVIDWebService}

\item{ensemblAttributes}{A vector of ensembl attributes}

\item{ensemblDataset}{Passed directly to biomaRt::useEnsembl}

\item{ensemblVersion}{Passed directly to biomaRt::useEnsembl}

\item{ensemblMirror}{Passed directly to biomaRt::useEnsembl}

\item{biomart}{Passed directly to biomaRt::useEnsembl}

\item{stringSpeciesId}{species ID. see Stringdb for list of avialable species}

\item{aliasPriorityOrder}{vector containig priority order of alias database. Must be UPPERCASE. Current databases: ENSEMBL, STRING, DAVID}
}
\description{
aliasTable
}
