% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{AppendCellHashing}
\alias{AppendCellHashing}
\title{AppendCellHashing}
\usage{
AppendCellHashing(seuratObj, barcodeCallFile, barcodePrefix = NULL)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
Appends cell hashing calls to a seurat object
}
