% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{appendTcrClonotypes}
\alias{appendTcrClonotypes}
\title{FunctionName.}
\usage{
appendTcrClonotypes(seuratObject = NA, clonotypeFile = NA,
  barcodePrefix = NULL)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
