% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{AvgCellExprs}
\alias{AvgCellExprs}
\title{AvgCellExprs}
\usage{
AvgCellExprs(
  seuratObj,
  varName = "ClusterNames_0.2",
  genes,
  slot = "data",
  useSeuratMethod = T
)
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{varName}{The resolution/ident to use}

\item{genes}{A vector of genes to include}

\item{slot}{The slot to use, passed to GetAssayData}

\item{useSeuratMethod}{This function now defaults to the base Seurat method set to T. If a simple average is needed set to F.}
}
\value{
A data.frame of avg expression per var
}
\description{
AvgCellExprs
}
