% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Sim.R
\name{bimodal_posDisHD}
\alias{bimodal_posDisHD}
\title{bimodal_posDisHD}
\usage{
bimodal_posDisHD(
  ndim = 5,
  n = 1000,
  cpct = 0.4,
  mu1 = log(1),
  mu2 = log(100),
  sig1 = log(3),
  sig2 = log(2),
  p = 0.5,
  JitterPeaks = T,
  removeDups = T,
  tSNEplot = T
)
}
\description{
ndim-HD bimodal Dist given parameters
}
