% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_SGS.R
\name{CalculateModuleScoreAvg}
\alias{CalculateModuleScoreAvg}
\title{CalculateModuleScoreAvg}
\usage{
CalculateModuleScoreAvg(seuratObj, genes.list = NULL,
  genes.pool = NULL, n.bin = 25, ctrl.size = 100, assay = NULL,
  seed.use = 1)
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{genes.list, }{A charater vector of genes to score}

\item{genes.pool, }{Gene list to base as the pool; NULL = all.}

\item{n.bin, }{number of bins to evaluate score across; default 25.}

\item{ctrl.size, }{control gene set size.}

\item{assay, }{The seurat assay to use.  Defaults to the result of DefaultAssay()}

\item{seed.use, }{random seed}
}
\value{
A dataframe with CellBarcode and Score for this gene set
}
\description{
Instead of the status quo score of Seurat AddModuleScore which is 1 score 1 gene, this function takes a list of genes and computes per set, the average of the individual scores.
}
