% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR.R
\name{CalculateTCRFreqForActivatedCells}
\alias{CalculateTCRFreqForActivatedCells}
\title{CalculateTCRFreqForActivatedCells}
\usage{
CalculateTCRFreqForActivatedCells(cDndIds,
  geneSetName = "HighlyActivated", positivityThreshold = 0.5,
  outPrefix = "./", invert = FALSE, doCleanup = FALSE)
}
\arguments{
\item{cDndIds}{A vector of cDNA rowIDs}

\item{geneSetName}{The gene set name to use for SGS}

\item{positivityThreshold}{The threshold to use for calling cells as positive}

\item{outPrefix}{A string that will be prepended to all saved files}

\item{invert}{If TRUE, those cells NOT positive for the gene set will be summarized, instead of positive cells}

\item{doCleanup}{If TRUE, any downloaded files will be deleted on completion}
}
\value{
A dataframe of results
}
\description{
For the supplied cDNA Rows, this will query their gene expression and TCR readsets, identifying a) existing seurat objects, b) vloupe files. If both are found,
}
