% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR.R
\name{CalculateTCRFreqForActivatedCellsAndImport}
\alias{CalculateTCRFreqForActivatedCellsAndImport}
\title{CalculateTCRFreqForActivatedCellsAndImport}
\usage{
CalculateTCRFreqForActivatedCellsAndImport(
  cDndIds,
  workbook = NULL,
  geneSetName = "HighlyActivated",
  positivityThreshold = 0.5,
  outPrefix = "./",
  assayName = "TCRdb",
  populationNameSuffix = "-HA",
  invert = FALSE,
  doCleanup = FALSE,
  minCells = NULL
)
}
\arguments{
\item{cDndIds}{A vector of cDNA rowIDs}

\item{workbook}{The target workbook to save results}

\item{geneSetName}{The gene set name to use for SGS}

\item{positivityThreshold}{The threshold to use for calling cells as positive}

\item{outPrefix}{A string that will be prepended to all saved files}

\item{assayName}{The name of the target assay}

\item{populationNameSuffix}{This string will be added to the end of the population field on the saved assay rows, appended to the associated cDNA population}

\item{invert}{If TRUE, those cells NOT positive for the gene set will be summarized, instead of positive cells}

\item{doCleanup}{If TRUE, any downloaded files will be deleted on completion}

\item{minCells}{If provided, data will only be imported if at least this many cells exist for the cDNA library}
}
\value{
Returns the object representation of the experiment batch (from Rlabkey::labkey.experiment.saveBatch)
}
\description{
For the supplied cDNA Rows, this will call CalculateTCRFreqForActivatedCells(), and also import the results into the provided assay
}
