% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{CellCycleScoring_SERIII}
\alias{CellCycleScoring_SERIII}
\title{FunctionName.}
\usage{
CellCycleScoring_SERIII(object, s.features, g2m.features,
  set.ident = FALSE)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
