% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{CheckDuplicatedCellNames}
\alias{CheckDuplicatedCellNames}
\title{CheckDuplicatedCellNames}
\usage{
CheckDuplicatedCellNames(object.list, stop = TRUE)
}
\arguments{
\item{object.list}{A lists of Seurat objects}

\item{stop}{Boolean that determines if stop() or print() is used when duplicates are found}
}
\description{
A variant on Seurat CheckDuplicatedCellNames that reports the top duplicated barcodes.  Mainly for debugging
}
