% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Seurat.R
\name{ClassifyCellsCustom}
\alias{ClassifyCellsCustom}
\title{A Title}
\usage{
ClassifyCellsCustom(classifierFile, ClassifierNames = NULL, testing.data,
  log10T = T, assay = NULL)
}
\description{
A description
}
