% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Seurat.R
\name{ClassifySeurat}
\alias{ClassifySeurat}
\title{ClassifySeurat}
\usage{
ClassifySeurat(seuratObj, trainedClassifiers = NULL,
  moduleScoreGeneLists = NULL, savePrefix = NULL,
  doGarnettClassify = F,
  garnett.path = "./data/Garnett/pbmc_classification.txt",
  assay = NULL)
}
\value{
A modified Seurat object.
}
\description{
This will classify cells in a seurat object using the specified classifiers
}
