% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_SGS.R
\name{ClassifySGS}
\alias{ClassifySGS}
\title{ClassifySGS}
\usage{
ClassifySGS(seuratObj, geneList = NULL, geneSetName = NULL,
  saveFilePath = NULL, doPlot = T, positivityThreshold = NULL)
}
\arguments{
\item{geneList}{A character vector of the genes to test}

\item{geneSetName}{A name to use for this gene list.  This will be used in the header of the dataframe}

\item{saveFilePath}{If provided, the dataframe will be saved here}

\item{doPlot}{If true, a QC plot will be created}

\item{positivityThreshold}{If provided, a boolean column will be added to the dataframe scoring cells as positive or negative, depending on if their score is greater than this threshold}
}
\value{
A dataframe with the results per cell
}
\description{
This will classify cells in a seurat object using a variant of Seurat's AddModuleScore
}
