% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_SGS.R
\name{ClassifySGS_Multiple}
\alias{ClassifySGS_Multiple}
\title{ClassifySGS_Multiple}
\usage{
ClassifySGS_Multiple(seuratObj, moduleScoreGeneLists = NULL,
  saveFilePath = NULL)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{moduleScoreGeneLists}{A named list of character vectors, where each vector is the set of genes to score}

\item{saveFilePath}{If provided, the datframe will be saved here}
}
\value{
A dataframe with the results per cell, for all gene sets
}
\description{
This will classify cells in a seurat object using a list of gene sets
}
