% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_SGS.R
\name{ClassifySGSAndApply}
\alias{ClassifySGSAndApply}
\title{ClassifySGSAndApply}
\usage{
ClassifySGSAndApply(
  seuratObj,
  geneList = NULL,
  geneSetName = NULL,
  saveFilePath = NULL,
  doPlot = T,
  positivityThreshold = NULL,
  reduction = NULL
)
}
\arguments{
\item{geneList}{A character vector of the genes to test}

\item{geneSetName}{A name to use for this gene list.  This will be used in the header of the dataframe}

\item{saveFilePath}{If provided, the dataframe will be saved here}

\item{doPlot}{If true, a QC plot will be created}

\item{positivityThreshold}{If provided, a boolean column will be added to the dataframe scoring cells as positive or negative, depending on if their score is greater than this threshold}

\item{reduction}{The reduction (i.e. tsne or umap) that will be used when plotting}
}
\value{
The modified seurat object
}
\description{
This will classify cells in a seurat object using a variant of Seurat's AddModuleScore and save the results to the seurat object
}
