% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{createExampleData}
\alias{createExampleData}
\title{A Title}
\usage{
createExampleData(nRow = 100, nCol = 10)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
