% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{createSeuratObj}
\alias{createSeuratObj}
\title{FunctionName.}
\usage{
createSeuratObj(seuratData = NA, project = NA, minFeatures = 25,
  minCells = 0)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
