% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{createSeuratObj_SERIII}
\alias{createSeuratObj_SERIII}
\title{Create a Seurat 3 object}
\usage{
createSeuratObj_SERIII(seuratData = NA, project = NA,
  minFeatures = 25, minCells = 0, MitoGenesPattern = "^MT-")
}
\arguments{
\item{seuratData, }{A Seurat input data from Read10X().}

\item{project, }{Sets the project name for the Seurat object.}

\item{minFeatures, }{Include cells where at least this many features are detected.}

\item{minCells, }{Include features detected in at least this many cells.}
}
\value{
A Seurat object with p.mito calculated.
}
\description{
Create Seurat Object from Read10X(). old name createSeuratObj.
}
\keyword{createSeuratObj}
