% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypingGeneSets.R
\name{CTL_Immune_GeneList}
\alias{CTL_Immune_GeneList}
\title{A Title}
\usage{
CTL_Immune_GeneList(QuickGO.path = "./data/QuickGO")
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
