% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{DoCellFiltering}
\alias{DoCellFiltering}
\title{A Title}
\usage{
DoCellFiltering(bData, minQuant = 0.05, maxValueForColSumFilter = 5)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
