% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellHashing.R
\name{doHtoDemux}
\alias{doHtoDemux}
\title{A Title}
\usage{
doHtoDemux(seuratObj)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
