% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQueries.R
\name{Download10xRawDataForLoupeFile}
\alias{Download10xRawDataForLoupeFile}
\title{Download10xRawDataForLoupeFile}
\usage{
Download10xRawDataForLoupeFile(
  outputFileId,
  outFile,
  overwrite = T,
  countType = "raw_feature_bc_matrix"
)
}
\arguments{
\item{outputFileId}{The outputfile Id of the loupe file}

\item{outFile}{The local path to write this file. The data will be written to a subfolder named raw_feature_bc_matrix}

\item{overwrite}{If true, any pre-existing local copy will be replaced.}

\item{countType}{Either raw_feature_bc_matrix or filtered_feature_bc_matrix}
}
\description{
Downloads the raw_feature_bc_matrix folder associated with the provided Loupe file
}
