% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{DownloadAndAppendCiteSeq}
\alias{DownloadAndAppendCiteSeq}
\title{DownloadAndAppendCiteSeq}
\usage{
DownloadAndAppendCiteSeq(
  seuratObj,
  outPath = ".",
  assayName = "ADT",
  featureLabelTable = NULL
)
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{assayName}{The name of the assay to store the ADT data.}

\item{featureLabelTable}{An optional CSV file listing the markers and metadata.}
}
\value{
A modified Seurat object.
}
\description{
Downloads matching Cite-seq counts using barcodePrefix on the seurat object
}
