% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{downloadAndAppendTcrClonotypes}
\alias{downloadAndAppendTcrClonotypes}
\title{A Title}
\usage{
downloadAndAppendTcrClonotypes(seuratObject, outPath = ".")
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
