% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{downloadCellRangerClonotypes}
\alias{downloadCellRangerClonotypes}
\title{A Title}
\usage{
downloadCellRangerClonotypes(vLoupeId, outFile, overwrite = T)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
