% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{DownloadOutputDirectoryFromOutputFile}
\alias{DownloadOutputDirectoryFromOutputFile}
\title{DownloadOutputFile}
\usage{
DownloadOutputDirectoryFromOutputFile(
  outputFileId,
  outFile,
  overwrite = T,
  pathTranslator = NULL
)
}
\arguments{
\item{outputFileId}{The rowid of the sequence outputfiles on the webserver}

\item{outFile}{The local path to write this file}

\item{overwrite}{If true, any pre-existing local copy will be replaced.}

\item{pathTranslator}{There are instances where the goal is to download some file located relative to the actual outputfile. For example, 10x data has an outputfile for the Loupe file, but it might be useful to download the raw counts. To accomplish this, a function can be provided that accepts the primary file remote path, and returns a modified path to download.}
}
\description{
Downloads an output file tracked in LabKey to the local filesystem.
}
