% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellHashing.R
\name{downloadOutputFile}
\alias{downloadOutputFile}
\title{A Title}
\usage{
downloadOutputFile(outputFileId, outFile, overwrite = T)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
