% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{ElbowPlot_SERIII}
\alias{ElbowPlot_SERIII}
\title{A Title}
\usage{
ElbowPlot_SERIII(object, ndims = 25, reduction = "pca",
  print.plot = T, return.elbow = T, min.y = 1.3)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
