% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz.R
\name{ExprThrTab}
\alias{ExprThrTab}
\title{PlotBin2D}
\usage{
ExprThrTab(SeuratObj, GeneName = NULL, CutThresh = NULL, PlotBar = T,
  MetaDataName = NULL, TitleExtra = "", PlotHist = F,
  col_vector = col_vector)
}
\arguments{
\item{SeuratObj, }{A Seurat object.}

\item{GeneName, }{the Gene name of interest.}

\item{CutThresh, }{cutting threshold.}

\item{TitleExtra, }{title from user.}
}
\value{
plot with ggplot tabulated bar plots.
}
\description{
This function from a seurat object, takes the gene expression of a single gene and provides a vizualization for thresholding and corresponding descritized tabulation.
}
