% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{Find_Markers}
\alias{Find_Markers}
\title{Find_Markers}
\usage{
Find_Markers(seuratObj, resolutionToUse, outFile, saveFileMarkers = NULL,
  testsToUse = c("wilcox", "bimod", "roc", "t", "negbinom", "poisson",
  "LR", "MAST", "DESeq2"), numGenesToSave = 20, onlypos = F)
}
\description{
Find_Markers
}
