% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{find_peaks}
\alias{find_peaks}
\title{find_peaks}
\usage{
find_peaks(x, m = 4)
}
\arguments{
\item{x, }{A vector of numbers, if small or not very smooth, use a smoothing function. Like density(x).}

\item{m, }{An integer that acts as a loose hand for resolution.}
}
\value{
vector of peaks positions.
}
\description{
Returns the maxima of points. To get minima, -1*x
}
