% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{findClustersAndDimRedux_SERIII}
\alias{findClustersAndDimRedux_SERIII}
\title{A Title}
\usage{
findClustersAndDimRedux_SERIII(seuratObj, dimsToUse = NULL,
  saveFile = NULL, forceReCalc = F, doUMAP = T)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
