% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R, R/mcl.R
\name{findElbow}
\alias{findElbow}
\title{A Title}
\usage{
findElbow(y, plot = FALSE, returnIndex = TRUE, ignore.concavity = F,
  min.y = NA, min.x = NA)

findElbow(y, plot = FALSE, returnIndex = TRUE, ignore.concavity = F,
  min.y = NA, min.x = NA)
}
\arguments{
\item{SeurObj, }{A Seurat object.}

\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.

A modified Seurat object.
}
\description{
A description

A description
}
\examples{
tmp <- findElbow(c(0.9, 1.1, 1.1, 1.9, 2.5, 2.8, 4.9, 8.5),
	plot = TRUE) # wandering
tmp <- findElbow(c(0.9, 1.0, 1.2, 1.3, 1.5, 1.5, 10.0, 22.0),
	plot = TRUE) # late rise
tmp <- findElbow(c(2, 4, 6, 8, 10, 12, 14, 16)^2,
	plot = TRUE) # gradual, no obvious break

# Not the usual way to choose the number of PCs:
library("chemometrics")
data(glass)
pca <- prcomp(glass)
eigensum <- sum(pca$sdev * pca$sdev)
vv <- 100 * (pca$sdev * pca$sdev/eigensum)
cs <- cumsum(vv)
tmp <- findElbow(vv, plot = TRUE)
tmp <- findElbow(cs, plot = TRUE)

tmp <- findElbow(c(0.9, 1.1, 1.1, 1.9, 2.5, 2.8, 4.9, 8.5),
	plot = TRUE) # wandering
tmp <- findElbow(c(0.9, 1.0, 1.2, 1.3, 1.5, 1.5, 10.0, 22.0),
	plot = TRUE) # late rise
tmp <- findElbow(c(2, 4, 6, 8, 10, 12, 14, 16)^2,
	plot = TRUE) # gradual, no obvious break

# Not the usual way to choose the number of PCs:
library("chemometrics")
data(glass)
pca <- prcomp(glass)
eigensum <- sum(pca$sdev * pca$sdev)
vv <- 100 * (pca$sdev * pca$sdev/eigensum)
cs <- cumsum(vv)
tmp <- findElbow(vv, plot = TRUE)
tmp <- findElbow(cs, plot = TRUE)

}
\keyword{SerIII_template}
