% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{findMarkers}
\alias{findMarkers}
\title{FunctionName.}
\usage{
findMarkers(seuratObj, resolutionToUse, outFile, saveFileMarkers = NULL,
  testsToUse = c("wilcox", "bimod", "roc", "t", "negbinom", "poisson",
  "LR", "MAST", "DESeq2"), numGenesToSave = 20)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
