% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{findMarkers_SERIII}
\alias{findMarkers_SERIII}
\title{A Title}
\usage{
findMarkers_SERIII(seuratObj, resolutionToUse, outFile,
  saveFileMarkers = NULL, testsToUse = c("wilcox", "bimod", "roc", "t",
  "negbinom", "poisson", "LR", "MAST", "DESeq2"), numGenesToSave = 20,
  onlypos = F)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
