% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{findSeuratElbow}
\alias{findSeuratElbow}
\title{FunctionName.}
\usage{
findSeuratElbow(seuratObj, ndims = 25, reduction = "pca",
  print.plot = T)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
