% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Garnett.R
\name{Garnett_Classification}
\alias{Garnett_Classification}
\title{A Title}
\usage{
Garnett_Classification(SerObj,
  marker_file_path = "./data/Garnett/pbmc_classification.txt",
  reutrnMonObj = F, returnTrainedClassifier = T)
}
\arguments{
\item{SerObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
