% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellHashing.R
\name{generateByRowSummary}
\alias{generateByRowSummary}
\title{A Title}
\usage{
generateByRowSummary(barcodeData)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
