% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateCellHashCallsMultiSeq}
\alias{GenerateCellHashCallsMultiSeq}
\title{A Title}
\usage{
GenerateCellHashCallsMultiSeq(barcodeData)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
