% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellHashing.R
\name{generateCellHashCallsSeurat}
\alias{generateCellHashCallsSeurat}
\title{A Title}
\usage{
generateCellHashCallsSeurat(barcodeData)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
