% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{GenerateSingleR}
\alias{GenerateSingleR}
\title{generate SingleR object}
\usage{
GenerateSingleR(seuratObj = NULL, dataset = "hpca", assay = NULL)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{dataset}{The dataset (see singleR docs) to use as a reference}

\item{assay}{The assay in the seurat object to use}
}
\value{
SingleR object
}
\description{
Compute SingleR classification on a Seurat object
}
\keyword{Classification}
\keyword{SingleR}
