% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellHashing.R
\name{generateSummaryForExpectedBarcodes}
\alias{generateSummaryForExpectedBarcodes}
\title{A Title}
\usage{
generateSummaryForExpectedBarcodes(dt, whitelistFile, outputFile,
  barcodeData)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
