% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQueries.R
\name{getCdnaRecords}
\alias{getCdnaRecords}
\title{A Title}
\usage{
getCdnaRecords(readsetId, type = "GEX")
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
