% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNames.R
\name{GetEnsembl}
\alias{GetEnsembl}
\title{GetEnsembl}
\usage{
GetEnsembl(seuratObj, MmulV = 8)
}
\arguments{
\item{seuratObj}{A Seurat Object}

\item{MmulV}{A numeric value 8 or 10}
}
\description{
GetEnsembl
}
