% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Caret.R
\name{GetEnsmbleVarImportance}
\alias{GetEnsmbleVarImportance}
\title{A Title}
\usage{
GetEnsmbleVarImportance(ClassiferSet, AvailableClassifiers = "",
  ScaleVarImp = T)
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
