% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{GetGeneIds}
\alias{GetGeneIds}
\title{Retrieve the gene IDs from a seuratObj created using ReadAndFilter10xData.}
\usage{
GetGeneIds(seuratObj, geneNames, throwIfGenesNotFound = TRUE)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{throwIfGenesNotFound}{If true and any of the requested gene names are not found, an error will be thrown.  Otherwise, the result will contain NAs}

\item{datasetName}{A name to use when creating the Seurat object}
}
\value{
A named vector of the gene IDs
}
\description{
Retrieve the gene IDs from a seuratObj created using ReadAndFilter10xData.
}
