% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Caret.R
\name{GetTraingStats}
\alias{GetTraingStats}
\title{A Title}
\usage{
GetTraingStats(ClassifierSet = NA, AvailableTrainConfMatrix = "",
  AvailableTestConfMatrix = "")
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
