% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellHashing.R
\name{HTODemux2}
\alias{HTODemux2}
\title{A Title}
\usage{
HTODemux2(object, assay = "HTO", positive.quantile = 0.99,
  nstarts = 100, kfunc = "clara", nsamples = 100, verbose = TRUE)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
