% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{HtoSummary}
\alias{HtoSummary}
\title{A Title}
\usage{
HtoSummary(seuratObj, field1, field2, doHeatmap = T)
}
\arguments{
\item{seurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
