% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{InferThresholds}
\alias{InferThresholds}
\title{A Title}
\usage{
InferThresholds(data, dataLabel, minQuant = 0.05, plotFigs = T,
  FindElbowMinY = NA)
}
\value{
A modified Seurat object.
}
\description{
A description
}
