% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{LengthCheck}
\alias{LengthCheck}
\title{A Title}
\usage{
LengthCheck(values, cutoff = 0)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
