% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{MakeSerObjs_10XFolders_SERIII}
\alias{MakeSerObjs_10XFolders_SERIII}
\title{A Title}
\usage{
MakeSerObjs_10XFolders_SERIII(counts.path = NULL, min.cells = 0,
  min.genes = 0, ProjName = "10X", save.path = NULL,
  returnList = F, path.exclude = "raw", string.exclude = NULL)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
