% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{markStepRun}
\alias{markStepRun}
\title{A Title}
\usage{
markStepRun(seuratObj, name, saveFile = NULL)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
