% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{MDSmyDF}
\alias{MDSmyDF}
\title{A Title}
\usage{
MDSmyDF(dfx, labelsDF, factorV, title = "MDS Plot", col_vector)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
