% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mmul10_conv1}
\alias{Mmul10_conv1}
\title{Table mapping ENS to SYM for MMul10}
\format{A data frame with 35427 rows and 2 variables:
\describe{
  \item{ENSID}{ENS ID}
  \item{SYMB}{Gene Symbol}
  ...
}}
\source{
\url{https://}
}
\usage{
Mmul10_conv1
}
\description{
Table mapping ENS to SYM for MMul10
}
\keyword{datasets}
