% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{PCAmyDF}
\alias{PCAmyDF}
\title{PCAmyDF}
\usage{
PCAmyDF(dfx, labels, factorV, title = "PCA Plot", scale, center,
  col_vector, namePointBL = F)
}
\value{
PCA plot of data or data
}
\description{
Compute PCA on a dataframe
}
\keyword{PCA,}
\keyword{ggplot2}
