% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{PerformEmptyDropletFiltering}
\alias{PerformEmptyDropletFiltering}
\title{PerformEmptyDropletFiltering}
\usage{
PerformEmptyDropletFiltering(seuratRawData, fdrThreshold = 0.01,
  emptyDropNIters = 10000)
}
\arguments{
\item{seuratRawData, }{raw data}
}
\value{
Plot
}
\description{
PerformEmptyDropletFiltering
}
