% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{PlotAvgExpr}
\alias{PlotAvgExpr}
\title{PlotAvgExpr}
\usage{
PlotAvgExpr(GenesNames2Show, X_avg, Y_avg, features = NULL,
  Xlab = "Xlab", Ylab = "Ylab", Title = "Title",
  HighColor = "dodgerblue")
}
\arguments{
\item{x, }{numbers.}
}
\value{
histo_numers
}
\description{

Plots average expression of each cluster/group for all associated cells

}
