% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz.R
\name{PlotBin2D}
\alias{PlotBin2D}
\title{PlotBin2D}
\usage{
PlotBin2D(DF, bins = 75, ncol = 3)
}
\arguments{
\item{DF, }{A dataframe with at least 3 columns, 2 for 2D scatter and one for the levels Named V1, V2, and Fac.}
}
\value{
plot with ggplot tabulated bar plots.
}
\description{
Plots a set of plots with geom_bin2d() split by factor levels
}
