% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiffusionMaps.R
\name{PlotDiffComp}
\alias{PlotDiffComp}
\title{Plots destiny's Diffusion Map}
\usage{
PlotDiffComp(diffMap, ndimsToPlot)
}
\arguments{
\item{ndimsToPlot}{number of dimensions to plot}

\item{DiffMap}{cell names}
}
\description{
Plots destiny's Diffusion Map
}
