% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiffusionMaps.R
\name{PlotDiffComp2}
\alias{PlotDiffComp2}
\title{Plots destiny's Diffusion Map. This function includes subplots that shows the trajectory of the cells by cell Ids on each componenets.}
\usage{
PlotDiffComp2(diffMap, ndimsToPlot)
}
\arguments{
\item{ndimsToPlot}{number of dimensions to plot}

\item{DiffMap}{cell names}
}
\description{
Plots destiny's Diffusion Map. This function includes subplots that shows the trajectory of the cells by cell Ids on each componenets.
}
