% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz.R
\name{PlotFeatThrTab}
\alias{PlotFeatThrTab}
\title{PlotFeatThrTab}
\usage{
PlotFeatThrTab(SeuratObj, FeatName = NULL, CutThresh = NULL,
  PlotBar = T, MetaDataName = NULL, TitleExtra = "", PlotHist = F,
  col_vector = col_vector)
}
\arguments{
\item{SeuratObj, }{A Seurat object.}

\item{FeatName, }{the Feature name of interest.}

\item{CutThresh, }{cutting threshold.}

\item{TitleExtra, }{title from user.}
}
\value{
plot with ggplot tabulated bar plots.
}
\description{
This function from a seurat object, takes the feature selected and provides a vizualization for thresholding and corresponding descritized tabulation.
}
