% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz.R
\name{PlotHistDensOptima}
\alias{PlotHistDensOptima}
\title{PlotHistDensOptima}
\usage{
PlotHistDensOptima(dftemp, Print = F, col_vector = col_vector,
  nn = NULL, Title = "")
}
\arguments{
\item{dftemp, }{a dataframe with column var1 to be plotted.}

\item{col_vector, }{color vector.}

\item{nn, }{a color from color_vector.}
}
\value{
plot with ggplot tabulated bar plots.
}
\description{
This function plots histogram+density and finds all optima from a df usually from FetchData() of Seurat.
}
