% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{PlotImmuneMarkers}
\alias{PlotImmuneMarkers}
\title{PlotImmuneMarkers}
\usage{
PlotImmuneMarkers(seuratObj, reduction = "tsne")
}
\arguments{
\item{reduction}{The reduction to use}
}
\description{
Generate a set of Seurat FeaturePlots for common immune cell markers
}
