% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz.R
\name{PlotMyTable}
\alias{PlotMyTable}
\title{PlotMyTable}
\usage{
PlotMyTable(MyTable, Title = "", legend.position = "bottom", PlotCombo = F)
}
\arguments{
\item{MyTable, }{A 2-way table e.g. MyTable = table(x, y)}
}
\value{
plot with ggplot tabulated bar plots.
}
\description{
Plots average expression of each cluster/group for all associated cells
}
