% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Seurat.R
\name{PredictSER}
\alias{PredictSER}
\title{A Title}
\usage{
PredictSER(ProcSERobj.path = NULL, PatternOfProcSERobj = "_proc.rds",
  classification.path = NULL, file.select = NULL,
  TrainedClassifiers.path = "../PBMC3k/data", save.fig.path = NULL,
  col_vector = NULL, returnLS = F, GarnettClassify = F,
  Garnett.path = "./data/Garnett/pbmc_classification.txt",
  MCEClassify = T, ModuleScoreGeneListClassify = F,
  ModuleScoreGeneLists = NULL,
  RhesusConvDavid.path = "./data/Rhesus/David6.8_ConvertedRhesus_ENSMMUG.txt",
  RhesusConvDavid = F, ENSMB.tag = "ENSMM", yhat.save.path = NA,
  cleanName = T, assay = NULL)
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
