% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{PreProcess_SerObjs}
\alias{PreProcess_SerObjs}
\title{PreProcess_SerObjs}
\usage{
PreProcess_SerObjs(SerObj.path = NULL, SerObjRDSKey = "SeuratObj.rds",
  ProjName = "10X", save.path = NULL, save.fig.path = NULL,
  returnList = F, returnDimReduxOnly = T, save.fig = T,
  nUMI.high = 20000, nGene.high = 3000, pMito.high = 0.15,
  nUMI.low = 0.99, nGene.low = 200, pMito.low = -Inf,
  fvg.x.low.cutoff = 0.01, fvg.x.high.cutoff = 4.5,
  fvg.y.cutoff = 1.5, KeepGene.LS = NULL, nDimPCA = NA,
  findPCAElbow = T, RemoveCellCycle = F,
  path2CCfiles = "./data/CellCycle", doUMAP = T)
}
\value{
A modified Seurat object.
}
\description{
PreProcess Seurat Objs
}
