% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollaboratorFxs_MyersGroup.R
\name{princurve_bootstrap}
\alias{princurve_bootstrap}
\title{princurve_bootstrap}
\usage{
princurve_bootstrap(input, df_range = NA, df_start = NA, df_end = NA,
  plot = TRUE)
}
\arguments{
\item{input, }{A matrix}
}
\value{
principal_curves.
}
\description{
A description
}
\details{
This function was originally found on 06/2019 in https://github.com/MyersGroup/testisAtlas.
}
\author{
Daniel Wells from Myers Group
}
\keyword{Psuedotime}
