% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{processAndAggregateTcrClonotypes}
\alias{processAndAggregateTcrClonotypes}
\title{FunctionName.}
\usage{
processAndAggregateTcrClonotypes(clonotypeFile)
}
\arguments{
\item{clonotypeFile, }{Expects all_contig_annotations.csv from cellranger vdj}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
