% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{ProcessCiteSeqCount}
\alias{ProcessCiteSeqCount}
\title{ProcessCiteSeqCount}
\usage{
ProcessCiteSeqCount(bFile = NA, doRowFilter = T, maxValueForColSumFilter = 5)
}
\arguments{
\item{bFile, }{The input barcode file}

\item{doRowFilter, }{If true, row-level (HTO) filtering will be performed}

\item{maxValueForColSumFilter, }{When column/cell filtering is performed, cell with low reads are dropped.  If this is provided, any column with at least this many reads will be retained.  Otherwise the threshold is determined automaticaly.  If NA, defaults to 5}
}
\value{

}
\description{
The primary entrypoint for demultiplexing HTO data
}
\keyword{CITE-Seq,}
