% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{ProcessCiteSeqData}
\alias{ProcessCiteSeqData}
\title{ProcessCiteSeqData}
\usage{
ProcessCiteSeqData(
  seuratObj,
  assayName = "ADT",
  dist.method = "euclidean",
  forceReCalc = F,
  print.plot = T,
  features = NULL
)
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{assayName}{The name of the assay to store the ADT data. default is 'ADT'}

\item{dist.method}{A distance method to compute SNN matrix}

\item{forceReCalc}{Force recalculate the pipeline}

\item{print.plot}{plotting bool}

\item{features}{features to run pipeline on}
}
\value{
A modified Seurat object.
}
\description{
Processes Cite-seq counts by running PCA, tSNE, and UMAP
}
