% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{ProcessEnsemblHtoCalls}
\alias{ProcessEnsemblHtoCalls}
\title{ProcessEnsemblHtoCalls}
\usage{
ProcessEnsemblHtoCalls(
  mc,
  sc,
  barcodeData,
  outFile = "combinedHtoCalls.txt",
  allCallsOutFile = NA
)
}
\value{
A modified Seurat object.
}
\description{
A description
}
