% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{processSeurat1}
\alias{processSeurat1}
\title{FunctionName.}
\usage{
processSeurat1(seuratObj, saveFile = NULL, doCellCycle = T,
  doCellFilter = F, nUMI.high = 20000, nGene.high = 3000,
  pMito.high = 0.15, nUMI.low = 0.99, nGene.low = 200,
  pMito.low = -Inf, forceReCalc = F, variableGeneTable = NULL,
  variableFeatureSelectionMethod = "vst")
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
