% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQueries.R
\name{QueryAndApplyCdnaMetadata}
\alias{QueryAndApplyCdnaMetadata}
\title{QueryAndApplyCdnaMetadata}
\usage{
QueryAndApplyCdnaMetadata(seuratObj, fieldSelect = c("rowid",
  "sortid/population", "sortid/stimid/animalId", "sortid/stimid/date",
  "sortid/stimid/stim"), fieldNames = c("cDNA_ID", "Population",
  "SubjectId", "SampleDate", "Stim"), overwriteExisting = F)
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{fieldSelect}{The set of fields to query}

\item{fieldNames}{The labels to use for the fields}
}
\value{
A modified Seurat object.
}
\description{
This will read the barcodePrefix from the seurat object, query the LK server and apply the appropriate metadata from the cDNA table
}
