% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNames.R
\name{QueryEnsemblSymbolAndHumanHomologs}
\alias{QueryEnsemblSymbolAndHumanHomologs}
\title{QueryEnsemblSymbolAndHumanHomologs}
\usage{
QueryEnsemblSymbolAndHumanHomologs(ensemblIds, biomart = "ensembl",
  dataset = "mmulatta_gene_ensembl",
  ensemblFilters = c("ensembl_gene_id"))
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs, passed to the getBM() filters argument}

\item{biomart}{Passed directly to biomaRt::useEnsembl}

\item{dataset}{Passed directly to biomaRt::useEnsembl}

\item{ensemblFilters}{A vector of ensembl IDs, passed to the getBM() filters argument.}
}
\description{
QueryEnsemblSymbolAndHumanHomologs
}
