% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{QuickSerCombObjs}
\alias{QuickSerCombObjs}
\title{QuickSerCombObjs}
\usage{
QuickSerCombObjs(save.fig.path = "./Figs",
  working.serobjs.path = "./data/10X/ser/proc", returnComboObj = F)
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
