% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{range01}
\alias{range01}
\title{A Title}
\usage{
range01(x, MaxN = NULL, MinN = NULL)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
