% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{range01b}
\alias{range01b}
\title{Range01b}
\usage{
range01b(x, MaxN = 10, MinN = -10)
}
\value{
A numeric vector
}
\description{
Scales the range of input to be between 0 and 1, w/ a prior Min Max
}
\keyword{range}
