% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{readAndFilter10xData}
\alias{readAndFilter10xData}
\title{Read and Filter 10X files}
\usage{
readAndFilter10xData(dataDir, datasetName)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
Reads in 10X files using Read10X and filters abberent cells using performEmptyDropletFiltering and returns a Seurat object
}
\examples{
readAndFilter10xData(dataDir="./10Xdata", datasetName = "my10X")
}
\keyword{ReadAndFilter10X}
