% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postSDA_processing.R
\name{RecomposeSDAmatrix}
\alias{RecomposeSDAmatrix}
\title{RecomposeSDAmatrix}
\usage{
RecomposeSDAmatrix(
  SDAseuratObj = NULL,
  WhichComp = "batch",
  metaDF = NULL,
  spikeGenes = NULL,
  nfeatures = 2000,
  tSNE_perplexity = 100,
  UMAP_MinDist = 0.5,
  UMAP_NumNeib = 60L,
  saveFile = NULL
)
}
\arguments{
\item{SDAseuratObj}{Final seurat object from SDA processing}

\item{WhichComp}{Three (3) options, "all", "qc", or "batch" default to "batch" which is comps removed by ShinySDA.}

\item{metaDF}{metadata datframe}

\item{spikeGenes}{list of genes to be added beyond the default VariableGenes()}

\item{nfeatures}{The number of VariableFeatures to identify}

\item{tSNE_perplexity}{tSNE perplexity. Passed directly to Seurat::RunTSNE()}

\item{UMAP_MinDist}{UMAP min. distance. Passed directly to Seurat::RunUMAP()}

\item{UMAP_NumNeib}{UMAP number of neighboring points. Passed directly to Seurat::RunUMAP()}

\item{saveFile}{If provided, the seurat object will be saved as RDS to this location}

\item{UMAP.NumEpoc}{UMAP min. distance. Passed directly to Seurat::RunUMAP()}
}
\description{
RecomposeSDAmatrix
}
