% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postSDA_processing.R
\name{RecreateSeuratObjFromSDAmatrix}
\alias{RecreateSeuratObjFromSDAmatrix}
\title{RecreateSeuratObjFromSDAmatrix}
\usage{
RecreateSeuratObjFromSDAmatrix(
  cellnames = NULL,
  recomposedMat = NULL,
  metaDF,
  projectName = "SDAproject",
  dimsToUse = 15,
  spikeGenes = NULL,
  nfeatures = 500,
  tSNE_perplexity = 0.3,
  UMAP_MinDist = 0.3,
  UMAP_NumNeib = 35L,
  UMAP.NumEpoc = 500
)
}
\arguments{
\item{cellnames}{cell names}

\item{recomposedMat}{recomposed matrix}

\item{metaDF}{metadata datframe}

\item{projectName}{Title of project}

\item{dimsToUse}{The number of dims to use}

\item{spikeGenes}{list of genes to be added beyond the default VariableGenes()}

\item{nfeatures}{The number of VariableFeatures to identify}

\item{tSNE_perplexity}{tSNE perplexity. Passed directly to Seurat::RunTSNE()}

\item{UMAP_MinDist}{UMAP min. distance. Passed directly to Seurat::RunUMAP()}

\item{UMAP_NumNeib}{UMAP number of neighboring points. Passed directly to Seurat::RunUMAP()}

\item{UMAP.NumEpoc}{UMAP min. distance. Passed directly to Seurat::RunUMAP()}
}
\description{
RecreateSeuratObjFromSDAmatrix
}
