% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypingGeneSets.R
\name{RhesusGeneDavidConv}
\alias{RhesusGeneDavidConv}
\title{A Title}
\usage{
RhesusGeneDavidConv(ColNames2Conv, RhesusConvDavid.path,
  ENSMB.tag = "ENSMM", returnFull = F)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
