% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{RunSingleR}
\alias{RunSingleR}
\title{generate SingleR object}
\usage{
RunSingleR(seuratObj = NULL, dataset = "hpca", assay = NULL,
  resultTableFile = NULL)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{dataset}{The dataset (see singleR docs) to use as a reference}

\item{assay}{The assay in the seurat object to use}

\item{resultTableFile}{If provided, a table of results will be saved here}
}
\value{
The modified seurat object
}
\description{
Compute SingleR classification on a Seurat object
}
\keyword{Classification}
\keyword{SingleR}
