% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{saveDimRedux}
\alias{saveDimRedux}
\title{FunctionName.}
\usage{
saveDimRedux(serObj, reductions = c("pca", "tsne", "umap"), file = NA,
  maxPCAcomps = 10)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
