% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Seurat.R, R/Seurat_III.R
\name{SaveDimRedux_SERIII}
\alias{SaveDimRedux_SERIII}
\title{A Title}
\usage{
SaveDimRedux_SERIII(seuratObj, reductions = c("pca", "tsne", "umap"),
  save.path = NA, maxPCAcomps = 10, nameID = "", returnResults = F,
  saveResults = T)

SaveDimRedux_SERIII(seuratObj, reductions = c("pca", "tsne", "umap"),
  save.path = NA, maxPCAcomps = 10, nameID = "", returnResults = F,
  saveResults = T)
}
\arguments{
\item{SeurObj, }{A Seurat object.}

\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.

A modified Seurat object.
}
\description{
A description

A description
}
\keyword{SerIII_template}
