% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{SaveSingleRtoSeuratObj}
\alias{SaveSingleRtoSeuratObj}
\title{Save SingleR to SeurObject}
\usage{
SaveSingleRtoSeuratObj(seuratObj, singler)
}
\value{
modified Seurat object
}
\description{
Add a pre-Computed SingleR classification into a Seurat object from path or direcly
}
\keyword{Classification}
\keyword{SingleR}
