% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{SaveSingleRtoSeurObj}
\alias{SaveSingleRtoSeurObj}
\title{Save SingleR to SeurObj}
\usage{
SaveSingleRtoSeurObj(SeurObjPath = NULL, SeurObj = NULL,
  SingleRPath = NULL)
}
\arguments{
\item{SeurObjPath, }{path to Seurat object}

\item{SeurObj}{a Seurat object, but if path given, path is prioritized.}

\item{SingleRPath}{path to SingleR object}
}
\value{
modified Seurat object
}
\description{
Add a pre-Computed SingleR classification into a Seurat object from path or direcly
}
\keyword{Classification}
\keyword{SingleR}
