% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monocle_III.R
\name{Ser2Monicle_MakeNProcess}
\alias{Ser2Monicle_MakeNProcess}
\title{A Title}
\usage{
Ser2Monicle_MakeNProcess(
  SeurObj = NULL,
  retunMon = T,
  PCAnDim = 20,
  doUMAP = T,
  min_dist = 0.3,
  n_neighbors = 40,
  dotSNE = F,
  doClust = T,
  ClusLouvRes = 5e-05,
  louvain_iter = 3,
  KeepTopNgenes = 3000,
  minExprGeneDet = 0.1,
  upperScale = 2,
  lowerScale = 2.5,
  assay = NULL
)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
