% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{SetIfNull}
\alias{SetIfNull}
\title{SetIfNull}
\usage{
SetIfNull(x, default)
}
\value{
A vector x
}
\description{
checks is.null, used in various Seurat based functions.
}
\keyword{Seurat}
\keyword{is.null,}
