% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{SingleRmySerObj}
\alias{SingleRmySerObj}
\title{SingleR my Seurat Object}
\usage{
SingleRmySerObj(SeurObjPath = NULL, SeurObj = NULL, PlotFigs = T,
  Annotation = NULL, ProjName = NULL, MinGenes = 500,
  Species = "Human", NumCores = 4, ClusteringName = NULL,
  SavePath = NULL)
}
\arguments{
\item{SeurObjPath, }{path to Seurat object}

\item{SeurObj}{a Seurat object, but if path given, path is prioritized.}
}
\value{
SingleR object
}
\description{
Compute SingleR classification on a Seurat object from path or direcly
}
\keyword{Classification}
\keyword{SingleR}
