% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{Tabulate_SingleRClassLabs}
\alias{Tabulate_SingleRClassLabs}
\title{Tabulate SingleR Class Lables}
\usage{
Tabulate_SingleRClassLabs(seuratObject)
}
\arguments{
\item{seuratObject}{a Seurat object, but if path given, path is prioritized.}
}
\description{
Tabulate Seurobject with SingleR class lables
}
\keyword{Classification}
\keyword{SingleR}
\keyword{Tabulate}
