% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAliasing.R
\name{TranslateGeneNames}
\alias{TranslateGeneNames}
\title{aliasTable}
\usage{
TranslateGeneNames(
  ensemblIds = NULL,
  geneSymbols = NULL,
  davidEmail,
  ensemblDataset = "mmulatta_gene_ensembl",
  ensemblVersion = NULL,
  ensemblMirror = "uswest",
  biomart = "ensembl",
  stringSpeciesId = 9606,
  useEnsembl = TRUE,
  useSTRINGdb = TRUE,
  useDAVID = TRUE
)
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs}

\item{geneSymbols}{A vector of gene symbols}

\item{davidEmail}{email account registered with DAVIDWebService}

\item{ensemblDataset}{Passed directly to biomaRt::useEnsembl}

\item{ensemblVersion}{Passed directly to biomaRt::useEnsembl}

\item{ensemblMirror}{Passed directly to biomaRt::useEnsembl}

\item{biomart}{Passed directly to biomaRt::useEnsembl}

\item{stringSpeciesId}{species ID. see Stringdb for list of available species}

\item{ensemblAttributes}{A vector of ensembl attributes}

\item{aliasPriorityOrder}{vector containig priority order of alias database. Must be UPPERCASE. Current databases: ENSEMBL, STRING, DAVID}
}
\description{
aliasTable
}
