% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAliasing.R
\name{TranslateToDAVID}
\alias{TranslateToDAVID}
\title{TranslateToDAVID}
\usage{
TranslateToDAVID(ensemblIds = NULL, geneSymbols = NULL, email)
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs}

\item{geneSymbols}{A vector of gene symbols}

\item{email}{email account registered with DAVIDWebService}
}
\description{
TranslateToDAVID
}
