% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAliasing.R
\name{TranslateToEnsembl}
\alias{TranslateToEnsembl}
\title{TranslateToEnsembl}
\usage{
TranslateToEnsembl(
  ensemblIds = NULL,
  geneSymbols = NULL,
  dataset = "mmulatta_gene_ensembl",
  ensemblVersion = NULL,
  ensemblMirror = "uswest",
  biomart = "ensembl"
)
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs, passed to the biomaRt::getBM() to query against ensembl_gene_id}

\item{geneSymbols}{A vector of gene symbols, passed to the biomaRt::getBM() to query against hgnc_symbol}

\item{dataset}{Passed directly to biomaRt::useEnsembl}

\item{ensemblVersion}{Passed directly to biomaRt::useEnsembl}

\item{ensemblMirror}{Passed directly to biomaRt::useEnsembl}

\item{biomart}{Passed directly to biomaRt::useEnsembl}
}
\description{
TranslateToEnsembl
}
