% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAliasing.R
\name{TranslateToStringDb}
\alias{TranslateToStringDb}
\title{TranslateToStringDb}
\usage{
TranslateToStringDb(ensemblIds = NULL, geneSymbols = NULL, speciesId = 9606)
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs}

\item{geneSymbols}{A vector of gene symbols}

\item{speciesId}{Species ID. see Stringdb reference for list of avialable species}

\item{replaceUnmatched}{Logical. If TRUE, removes NA's and replaces with inputs}
}
\description{
TranslateToStringDb
}
