% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{TransposeDT}
\alias{TransposeDT}
\title{transposedt}
\usage{
TransposeDT(dt, varlabel = "myVar")
}
\arguments{
\item{dt}{The datatable}
}
\value{
A transposed data.table
}
\description{
Transpose a data.table
}
\keyword{t}
\keyword{transpose,}
