% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{UniformSampleDF_FacPor}
\alias{UniformSampleDF_FacPor}
\title{A Title}
\usage{
UniformSampleDF_FacPor(x, ClassF, p)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
