% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{WichIn1not2}
\alias{WichIn1not2}
\title{WichIn1not2}
\usage{
WichIn1not2(Clus1N = c(1), DataT = "", Clus2N = c(2))
}
\value{
unique genes to the comparison
}
\description{
compares unique association between two labels from A dataftrame usually from GO annotation with a column named cluster
}
\keyword{cluster,}
\keyword{gene,}
