% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{WilcoxDETest}
\alias{WilcoxDETest}
\title{A Title}
\usage{
WilcoxDETest(object, cells.1, cells.2, verbose = TRUE, ...)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
