% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{WriteSummaryMetrics}
\alias{WriteSummaryMetrics}
\title{WriteSummaryMetrics}
\usage{
WriteSummaryMetrics(seuratObj, file)
}
\arguments{
\item{file}{The file where metrics will be written}

\item{seurObj, }{A Seurat object.}
}
\description{
WriteSummaryMetrics
}
