% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{AddModuleScore_SERIII}
\alias{AddModuleScore_SERIII}
\title{FunctionName.}
\usage{
AddModuleScore_SERIII(object, genes.list = NULL, genes.pool = NULL,
  n.bin = 25, seed.use = 1, ctrl.size = 100, use.k = FALSE,
  enrich.name = "Cluster", random.seed = 1)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
