% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{addTitleToMultiPlot}
\alias{addTitleToMultiPlot}
\title{A Title}
\usage{
addTitleToMultiPlot(plotGrid, title, relHeights = c(0.1, 1))
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
