% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{AddTitleToMultiPlot}
\alias{AddTitleToMultiPlot}
\title{A Title}
\usage{
AddTitleToMultiPlot(plotGrid, title, relHeights = c(0.1, 1))
}
\arguments{
\item{plotGrid}{The list of plots}

\item{title}{The title for this plot}

\item{relHeights}{The relative heights, passed to plot_grid}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
