% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellHashing.R
\name{appendCellHashing}
\alias{appendCellHashing}
\title{A Title}
\usage{
appendCellHashing(seuratObj, barcodeCallFile, barcodePrefix = NULL)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
