% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{AppendTcrClonotypes}
\alias{AppendTcrClonotypes}
\title{AppendTcrClonotypes}
\usage{
AppendTcrClonotypes(seuratObject = NA, clonotypeFile = NA,
  barcodePrefix = NULL, dropExisting = F, metaFeat = NULL)
}
\description{
AppendTcrClonotypes data from Prime-Seq
}
