% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{appendTcrClonotypes}
\alias{appendTcrClonotypes}
\title{A Title}
\usage{
appendTcrClonotypes(seuratObject = NA, clonotypeFile = NA,
  barcodePrefix = NULL)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
