% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{AvgCellExprs}
\alias{AvgCellExprs}
\title{AvgCellExprs}
\usage{
AvgCellExprs(seuratObj, varName = "ClusterNames_0.2", Genes)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A data.frame of avg expression per var
}
\description{
AvgCellExprs
}
