% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{CellCycleScoring_SERIII}
\alias{CellCycleScoring_SERIII}
\title{A Title}
\usage{
CellCycleScoring_SERIII(object, s.features, g2m.features,
  set.ident = FALSE)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
