% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Caret.R
\name{ClassifyCellsCustom}
\alias{ClassifyCellsCustom}
\title{A Title}
\usage{
ClassifyCellsCustom(Classifier.rds.path = "", ClassifierNames = "",
  testing.data, log10T = T, returnTraining = F)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
