% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{createExampleData}
\alias{createExampleData}
\title{FunctionName.}
\usage{
createExampleData(nRow = 100, nCol = 10)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
