% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{DimPlot_SingleRClassLabs}
\alias{DimPlot_SingleRClassLabs}
\title{DimPlot SingleR Class Lables}
\usage{
DimPlot_SingleRClassLabs(seuratObject)
}
\arguments{
\item{seuratObject}{a Seurat object, but if path given, path is prioritized.}
}
\description{
Dimplot Seurobject with SingleR class lables
}
\keyword{Classification}
\keyword{Dimplot}
\keyword{SingleR}
