% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellHashing.R
\name{doCellFiltering}
\alias{doCellFiltering}
\title{A Title}
\usage{
doCellFiltering(bData, minQuant = 0.05, maxValueForColSumFilter = 5)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
