% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{DoHtoDemux}
\alias{DoHtoDemux}
\title{A Title}
\usage{
DoHtoDemux(seuratObj, positive.quantile = 0.99)
}
\arguments{
\item{seurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
