% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{doMergeCCA}
\alias{doMergeCCA}
\title{doMergeCCA}
\usage{
doMergeCCA(seuratObjs, nameList, maxCCAspaceDim, assay = NULL,
  normalization.method = "LogNormalize", useAllFeatures = F,
  includeCellCycleGenes = T, nVariableFeatures = NULL, plotFigs = T,
  spike.genes = NULL, maxPCs2Weight = NULL)
}
\arguments{
\item{seuratObjs}{A list of seurat objects}

\item{nameList}{A list of names}

\item{maxCCAspaceDim}{The number of dims to use with FindIntegrationAnchors()}

\item{assay}{NULL for default assay or specify}

\item{useAllFeatures}{If true, the resulting object will contain all features, as opposed to just VariableGenes (not recommended)}

\item{includeCellCycleGenes}{If true, the cell cycles genes will always be included with IntegrateData(), as opposed to just VariableGenes}

\item{nVariableFeatures}{The number of VariableFeatures to identify}

\item{spike.genes}{If NULL ignored, but a vector of unique genes}

\item{maxPCs2Weight}{The number of dims to use with IntegrateData()}
}
\value{
A modified Seurat object.
}
\description{
An internal method to do CCA merging
}
