% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{DoMULTIseqDemux}
\alias{DoMULTIseqDemux}
\title{A Title}
\usage{
DoMULTIseqDemux(seuratObj)
}
\arguments{
\item{seurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
