% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellHashing.R
\name{doRowFiltering}
\alias{doRowFiltering}
\title{doRowFiltering}
\usage{
doRowFiltering(bData, minRowSum = 5, minRowMax = 20,
  minRowMean = 0.2, minMeanNonZeroCount = 2)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
