% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{DoRowFiltering}
\alias{DoRowFiltering}
\title{DoRowFiltering}
\usage{
DoRowFiltering(
  bData,
  minRowSum = 5,
  minRowMax = 20,
  minRowMean = 0.2,
  minMeanNonZeroCount = 2
)
}
\value{
A modified Seurat object.
}
\description{
A description
}
