% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{DownloadAndAppendCiteSeq}
\alias{DownloadAndAppendCiteSeq}
\title{DownloadAndAppendCiteSeq}
\usage{
DownloadAndAppendCiteSeq(
  seuratObj,
  outPath = ".",
  assayName = "ADT",
  featureLabelTable = NULL,
  adtWhitelist = NULL
)
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{assayName}{The name of the assay to store the ADT data.}

\item{featureLabelTable}{An optional TSV file listing the markers and metadata.  It must contain a header row with at least the columns (lowercase): tagname, sequence. If it contains the column markername, this will be used to replace the rownames of the matrix.}

\item{adtWhitelist}{An optional character vector of tag names, which must exactly match the rownames of the ADT count matrix.  If provided, data will be limited to these tags.}
}
\value{
A modified Seurat object.
}
\description{
Downloads matching Cite-seq counts using barcodePrefix on the seurat object
}
