% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{DownloadAndAppendCiteSeq}
\alias{DownloadAndAppendCiteSeq}
\title{DownloadAndAppendCiteSeq}
\usage{
DownloadAndAppendCiteSeq(seuratObject, outPath = ".")
}
\arguments{
\item{seuratObject, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
Downloads matching Cite-seq counts using barcodePrefix on the seurat object
}
