% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQueries.R
\name{DownloadAndAppendTcrClonotypes}
\alias{DownloadAndAppendTcrClonotypes}
\title{DownloadAndAppendTcrClonotypes}
\usage{
DownloadAndAppendTcrClonotypes(seuratObject, outPath = ".",
  dropExisting = T)
}
\arguments{
\item{seuratObject}{A Seurat object}

\item{outPath}{The output filepath}

\item{dropExisting}{If true, any existing clonotype data will be replaced}
}
\value{
A modified Seurat object.
}
\description{
Download And Append TCR Clonotypes data from Prime-Seq
}
