% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{downloadAndAppendTcrClonotypes}
\alias{downloadAndAppendTcrClonotypes}
\title{FunctionName.}
\usage{
downloadAndAppendTcrClonotypes(seuratObject, outPath = ".")
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
