% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{DownloadAndAppendTcrClonotypes}
\alias{DownloadAndAppendTcrClonotypes}
\title{A Title}
\usage{
DownloadAndAppendTcrClonotypes(seuratObject, outPath = ".",
  dropExisting = T)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
