% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{DownloadCellRangerClonotypes}
\alias{DownloadCellRangerClonotypes}
\title{DownloadCellRangerClonotypes}
\usage{
DownloadCellRangerClonotypes(vLoupeId, outFile, overwrite = T)
}
\description{
Download CellRanger Clonotypes
}
