% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{downloadCellRangerClonotypes}
\alias{downloadCellRangerClonotypes}
\title{FunctionName.}
\usage{
downloadCellRangerClonotypes(vLoupeId, outFile, overwrite = T)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
