% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{DownloadOutputFile}
\alias{DownloadOutputFile}
\title{DownloadOutputFile}
\usage{
DownloadOutputFile(outputFileId, outFile, overwrite = T)
}
\arguments{
\item{outputFileId}{The rowid of the sequence outputfiles on the webserver}

\item{outFile}{The local path to write this file}

\item{overwrite}{If true, any pre-existing local copy will be replaced.}
}
\description{
Downloads an output file tracked in LabKey to the local filesystem.
}
