% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{FindClustersAndDimRedux}
\alias{FindClustersAndDimRedux}
\title{FindClustersAndDimRedux}
\usage{
FindClustersAndDimRedux(seuratObj, dimsToUse = NULL, saveFile = NULL,
  forceReCalc = F)
}
\arguments{
\item{seurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FindClustersAndDimRedux
}
