% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{findClustersAndDimRedux}
\alias{findClustersAndDimRedux}
\title{FunctionName.}
\usage{
findClustersAndDimRedux(seuratObj, dimsToUse = NULL, saveFile = NULL,
  forceReCalc = F)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
