% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{FindClustersAndDimRedux}
\alias{FindClustersAndDimRedux}
\title{FindClustersAndDimRedux}
\usage{
FindClustersAndDimRedux(
  seuratObj,
  dimsToUse = NULL,
  saveFile = NULL,
  forceReCalc = F,
  minDimsToUse = NULL,
  umap.method = "umap-learn",
  UMAP_NumNeib = 40L,
  UMAP_MinDist = 0.2,
  UMAP_Seed = 1234,
  UMAP.NumEpoc = 500,
  maxTsneIter = 10000,
  tSNE_perplexity = 30,
  clusterResSet = c(0.2, 0.4, 0.6, 0.8, 1.2)
)
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{dimsToUse}{The number of dims to use.  If null, this will be inferred using FindSeuratElbow()}

\item{saveFile}{If provided, the seurat object will be saved as RDS to this location}

\item{forceReCalc}{If true, all steps will be performed even if already marked complete}

\item{minDimsToUse}{The minimum numbers of dims to use.  If dimsToUse is provided, this will override.}

\item{umap.method}{The UMAP method, either uwot or umap-learn, passed directly to Seurat::RunUMAP}

\item{maxTsneIter}{The value of max_iter to provide to RunTSNE.  Increasing can help large datasets.}

\item{tSNE_perplexity}{tSNE perplexity. Passed directly to Seurat::RunTSNE(), but OOSAP:::.InferPerplexityFromSeuratObj() corrects it if need be based dataset dims.}

\item{clusterResSet}{a vector of clustering resolutions, default is (0.2, 0.4, 0.6, 0.8, 1.2).}
}
\value{
A modified Seurat object.
}
\description{
FindClustersAndDimRedux
}
