% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{FindClustersAndDimRedux}
\alias{FindClustersAndDimRedux}
\title{FindClustersAndDimRedux}
\usage{
FindClustersAndDimRedux(seuratObj, dimsToUse = NULL, saveFile = NULL,
  forceReCalc = F, UMAP_NumNeib = 40L, UMAP_MinDist = 0.2,
  UMAP_Seed = 1234, UMAP.NumEpoc = 500)
}
\arguments{
\item{seurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FindClustersAndDimRedux
}
