% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{FindClustersAndDimRedux}
\alias{FindClustersAndDimRedux}
\title{FindClustersAndDimRedux}
\usage{
FindClustersAndDimRedux(seuratObj, dimsToUse = NULL, saveFile = NULL,
  forceReCalc = F, minDimsToUse = NULL, umap.method = "umap-learn")
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{dimsToUse}{The number of dims to use.  If null, this will be inferred using FindSeuratElbow()}

\item{saveFile}{If provided, the seurat object will be saved as RDS to this location}

\item{forceReCalc}{If true, all steps will be performed even if already marked complete}

\item{minDimsToUse}{The minimum numbers of dims to use.  If dimsToUse is provided, this will override.}

\item{umap.method}{The UMAP method, either uwot or umap-learn, passed directly to Seurat::RunUMAP}

}
\value{
A modified Seurat object.
}
\description{
FindClustersAndDimRedux
}
