% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateCellHashCallsSeurat}
\alias{GenerateCellHashCallsSeurat}
\title{A Title}
\usage{
<<<<<<< HEAD
GenerateCellHashCallsSeurat(barcodeData, positive.quantile = 0.99)
=======
GenerateCellHashCallsSeurat(barcodeData, positive.quantile = 0.99,
  attemptRecovery = TRUE)
>>>>>>> 83d4652b13485717ae4d00f543eed4bb745e281d
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
