% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateCellHashCallsSeurat}
\alias{GenerateCellHashCallsSeurat}
\title{A Title}
\usage{
GenerateCellHashCallsSeurat(barcodeData, positive.quantile = 0.99)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
