% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{generateSingleR}
\alias{generateSingleR}
\title{generate SingleR object}
\usage{
generateSingleR(SeurObj = NULL, Annotation = Annotation,
  ProjName = ProjName, MinGenes = MinGenes, Species = Species,
  ClusteringName = ClusteringName, NumCores = NumCores,
  NormGeneLength = F, VarGeneMethod = "de", FineTune = T,
  DoSig = T, MainTypes = T, RedFS = T)
}
\arguments{
\item{SeurObj}{a Seurat object, but if path given, path is prioritized.}

\item{SeurObjPath, }{path to Seurat object}
}
\value{
SingleR object or Seurat object
}
\description{
Compute SingleR classification on a Seurat object
}
\keyword{Classification}
\keyword{SingleR}
