% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Caret.R
\name{GetEnsmbleVarImportance}
\alias{GetEnsmbleVarImportance}
\title{GetEnsmbleVarImportance}
\usage{
GetEnsmbleVarImportance(ClassiferSet, AvailableClassifiers = "",
  ScaleVarImp = T)
}
\value{
A list
}
\description{
Given a list of classifiers, retyrb a df of VarImportance
}
\keyword{caret}
