% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{ggUMAP}
\alias{ggUMAP}
\title{ggUMAP}
\usage{
ggUMAP(
  object,
  colFac = NULL,
  col_vector = NULL,
  ptSize = 0.15,
  ptAlpha = 0.5,
  add2title = "",
  legendTitle = "",
  cells.use = NULL
)
}
\arguments{
\item{object}{A seurat object}
}
\description{
Generated a ggplot object from the UMAP data of a seurat object
}
