% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{MDSmyDF}
\alias{MDSmyDF}
\title{MDSmyDF}
\usage{
MDSmyDF(dfx, labelsDF, factorV, title = "MDS Plot", col_vector,
  returnMDS = F)
}
\value{
MDS plot of data or data
}
\description{
Compute Multidimensional scaling on a dataframe
}
\keyword{MDS,}
\keyword{ggplot2}
