% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{mergeSeuratObjs}
\alias{mergeSeuratObjs}
\title{FunctionName.}
\usage{
mergeSeuratObjs(seuratObjs, data, alignData = T, MaxCCAspaceDim = 20,
  MaxPCs2Weight = 20)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
