% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{MergeSeuratObjs}
\alias{MergeSeuratObjs}
\title{A Title}
\usage{
MergeSeuratObjs(seuratObjs, metadata = NULL, alignData = T,
  MaxCCAspaceDim = 20, MaxPCs2Weight = 20, projectName = NULL,
  PreProcSeur = F)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
