% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{MergeSeuratObjs}
\alias{MergeSeuratObjs}
\title{MergeSeuratObjs}
\usage{
MergeSeuratObjs(seuratObjs, metadata = NULL, alignData = T,
  maxCCAspaceDim = 20, maxPCs2Weight = 20, projectName = NULL,
  doScaleEach = T, useAllFeatures = F, nVariableFeatures = 2000,
  includeCellCycleGenes = T, selection.method = "vst",
  SpikeInGenes = NULL)
}
\arguments{
\item{seuratObjs}{A list of seurat objects, optionally named (in which case these will be used as dataset names)}

\item{metadata}{A list of metadata.  If provided, the names of this list will be used as dataset names}

\item{alignData}{If true, data will be aligned using Seurat::IntegrateData()}

\item{maxCCAspaceDim}{The number of dims to use with FindIntegrationAnchors()}

\item{maxPCs2Weight}{The number of dims to use with IntegrateData()}

\item{projectName}{The project name when creating the final seuratObj}

\item{doScaleEach}{If true, scale=T will be passed to FindIntegrationAnchors(); its default behavior to find anchors.}

\item{useAllFeatures}{If true, the resulting object will contain all features, as opposed to just VariableGenes (not recommended)}

\item{nVariableFeatures}{If NULL, top 2000 genes are used. Default for speed is 2000, the number of VariableFeatures to identify}

\item{includeCellCycleGenes}{If true, the cell cycles genes will always be included with IntegrateData(), as opposed to just VariableGenes}

\item{selection.method}{can be vst, mvp, or disp. see selection.method paramter with ?Seurat::FindVariableFeatures for more details.}

\item{SpikeInGenes}{A vector of gene names in the original seurat object. Check %in% seuratObjs before using.}
}
\value{
A modified Seurat object.
}
\description{
Merges a list of Seurat objects, using Seurat::IntegrateData()
}
