% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{MergeSeuratObjs}
\alias{MergeSeuratObjs}
\title{MergeSeuratObjs}
\usage{
MergeSeuratObjs(seuratObjs, metadata = NULL, alignData = T,
  MaxCCAspaceDim = 20, MaxPCs2Weight = 20, projectName = NULL,
  PreProcSeur = F, useAllFeatures = F, nVariableFeatures = 2000,
  includeCellCycleGenes = T)
}
\value{
A modified Seurat object.
}
\description{
Merges a list of Seurat objects.
}
\keyword{merge}
