% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{mergeSeuratObjs}
\alias{mergeSeuratObjs}
\title{A Title}
\usage{
mergeSeuratObjs(seuratObjs, data, alignData = T, MaxCCAspaceDim = 20,
  MaxPCs2Weight = 20)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
