% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{PCAmyDF}
\alias{PCAmyDF}
\title{A Title}
\usage{
PCAmyDF(dfx, labels, factorV, title = "PCA Plot", scale, center,
  col_vector, namePointBL = F)
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
