% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{performEmptyDropletFiltering}
\alias{performEmptyDropletFiltering}
\title{FunctionName.}
\usage{
performEmptyDropletFiltering(seuratRawData, fdrThreshold = 0.01,
  emptyDropNIters = 10000)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
