% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{PerformEmptyDropletFiltering}
\alias{PerformEmptyDropletFiltering}
\title{PerformEmptyDropletFiltering}
\usage{
PerformEmptyDropletFiltering(seuratRawData, fdrThreshold = 0.01,
  emptyDropNIters = 10000)
}
\arguments{
\item{seuratRawData}{Raw data}

\item{fdrThreshold}{FDR threshold, passed directly to PerformEmptyDrops()}

\item{emptyDropNIters}{Number of iterations, passed directly to PerformEmptyDrops()}
}
\value{
Plot
}
\description{
PerformEmptyDropletFiltering
}
