% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{PerformEmptyDrops}
\alias{PerformEmptyDrops}
\title{PerformEmptyDrops}
\usage{
PerformEmptyDrops(seuratRawData, emptyDropNIters, fdrThreshold = 0.01)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
PerformEmptyDrops
}
