% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{performEmptyDrops}
\alias{performEmptyDrops}
\title{FunctionName.}
\usage{
performEmptyDrops(seuratRawData, emptyDropNIters, fdrThreshold = 0.01)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
