% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{processAndAggregateTcrClonotypes}
\alias{processAndAggregateTcrClonotypes}
\title{A Title}
\usage{
processAndAggregateTcrClonotypes(clonotypeFile)
}
\arguments{
\item{clonotypeFile, }{Expects all_contig_annotations.csv from cellranger vdj}

\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
