% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{ProcessCiteSeqCount}
\alias{ProcessCiteSeqCount}
\title{ProcessCiteSeqCount}
\usage{
ProcessCiteSeqCount(bFile = NA, doRowFilter = T, doColumnFilter = T)
}
\arguments{
\item{bFile, }{The input barcode file}

\item{doRowFilter, }{If true, row-level (HTO) filtering will be performed}

\item{doColumnFilter, }{If true column-level (cell) filtering will be performed}
}
\value{

}
\description{
The primary entrypoint for demultiplexing HTO data
}
\keyword{CITE-Seq,}
