% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{ProcessSeurat1}
\alias{ProcessSeurat1}
\title{A Title}
\usage{
ProcessSeurat1(seuratObj, saveFile = NULL, doCellCycle = T,
  doCellFilter = F, nUMI.high = 20000, nGene.high = 3000,
  pMito.high = 0.15, nUMI.low = 0.99, nGene.low = 200,
  pMito.low = -Inf, forceReCalc = F, variableGeneTable = NULL,
  variableFeatureSelectionMethod = "vst", nVariableFeatures = 2000,
  printDefaultPlots = T, npcs = 50)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
