% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQueries.R
\name{QueryAndApplyCdnaMetadata}
\alias{QueryAndApplyCdnaMetadata}
\title{QueryAndApplyCdnaMetadata}
\usage{
QueryAndApplyCdnaMetadata(seuratObj, fieldSelect = c("rowid",
  "sortid/population", "sortid/stimid/animalId", "sortid/stimid/date",
  "sortid/stimid/stim"), fieldNames = c("cDNA ID", "Population",
  "SubjectId", "SampleDate", "Stim"), overwriteExisting = F)
}
\arguments{
\item{seuratObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
QueryAndApplyCdnaMetadata
}
