% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNames.R
\name{QueryEnsemblSymbolAndHumanHomologs}
\alias{QueryEnsemblSymbolAndHumanHomologs}
\title{QueryEnsemblSymbolAndHumanHomologs}
\usage{
QueryEnsemblSymbolAndHumanHomologs(
  ensemblIds,
  biomart = "ensembl",
  dataset = "mmulatta_gene_ensembl",
  ensemblFilters = c("ensembl_gene_id"),
  version = NULL,
  ensemblMirror = NULL,
  extraAttrs = NULL
)
}
\arguments{
\item{ensemblIds}{A vector of ensembl IDs, passed to the getBM() filters argument}

\item{biomart}{Passed directly to biomaRt::useEnsembl}

\item{dataset}{Passed directly to biomaRt::useEnsembl}

\item{ensemblFilters}{A vector of ensembl IDs, passed to the getBM() filters argument.}

\item{version}{Passed directly to biomaRt::useEnsembl}

\item{ensemblMirror}{Passed directly to biomaRt::useEnsembl.  Defaults to uswest since the default (NULL) was given SSL errors in docker}

\item{extraAttrs}{A vector of attributes that biomaRt::getBM() should get beyond default set. see listAttributes.}
}
\description{
QueryEnsemblSymbolAndHumanHomologs
}
