% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{range01b}
\alias{range01b}
\title{A Title}
\usage{
range01b(x, MaxN = 10, MinN = -10)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
