% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{ReadAndFilter10xData}
\alias{ReadAndFilter10xData}
\title{Read and Filter 10X files.}
\usage{
ReadAndFilter10xData(dataDir, datasetName)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
Reads in 10X files using Read10X and filters abberent cells using PerformEmptyDropletFiltering and returns a Seurat object.
}
\keyword{ReadAndFilter10X}
