% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{RemoveCellCycle}
\alias{RemoveCellCycle}
\title{RemoveCellCycle}
\usage{
RemoveCellCycle(seuratObj, runPCAonVariableGenes = F)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
RemoveCellCycle
}
