% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{RemoveCellCycle}
\alias{RemoveCellCycle}
\title{RemoveCellCycle}
\usage{
RemoveCellCycle(
  seuratObj,
  pcaResultFile = NULL,
  useSCTransform = F,
  do.scale = T,
  do.center = T
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{pcaResultFile}{If provided, cell cycle PCA results will be written here}

\item{useSCTransform}{If true, SCTransform will be used in place of the standard Seurat workflow (NormalizeData, ScaleData, FindVariableFeatures)}

\item{do.scale}{If true, scale residuals to have unit variance; SCTransform default = F; ScaldData default = T}

\item{do.center}{If true,center residuals to have mean zero; SCTransform default = T; ScaldData default = T}
}
\value{
A modified Seurat object.
}
\description{
RemoveCellCycle
}
