% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{RunSingleR}
\alias{RunSingleR}
\title{generate SingleR object}
\usage{
RunSingleR(
  seuratObj = NULL,
  dataset = "hpca",
  assay = NULL,
  resultTableFile = NULL,
  singlerSavePrefix = NULL,
  minFraction = 0.01,
  showHeatmap = TRUE,
  maxCellsForHeatmap = 20000
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{dataset}{The dataset (see singleR docs) to use as a reference}

\item{assay}{The assay in the seurat object to use}

\item{resultTableFile}{If provided, a table of results will be saved here}

\item{singlerSavePrefix}{If provided, the SingleR}

\item{minFraction}{If provided, any labels present with fraction of this or fewer across cells will be converted to Unknown}

\item{showHeatmap}{If true, heatmaps will be generated showing the SingleR calls}

\item{maxCellsForHeatmap}{The heatmap will only be plotted if the total cells is below this number}
}
\value{
The modified seurat object
}
\description{
Compute SingleR classification on a Seurat object
}
\keyword{Classification}
\keyword{SingleR}
