% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Seurat.R, R/Seurat_III.R
\name{SaveDimRedux}
\alias{SaveDimRedux}
\title{A Title}
\usage{
SaveDimRedux(seuratObj, reductions = c("pca", "tsne", "umap"),
  file = NA, maxPCAcomps = 10, returnResults = F)

SaveDimRedux(seuratObj, reductions = c("pca", "tsne", "umap"),
  file = NA, maxPCAcomps = 10, returnResults = F)
}
\arguments{
\item{SeurObj, }{A Seurat object.}

\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.

A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
