% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{SaveDimRedux}
\alias{SaveDimRedux}
\title{SaveDimRedux}
\usage{
SaveDimRedux(seuratObj, reductions = c("pca", "tsne", "umap"),
  file = NA, maxPCAcomps = 10, returnResults = F)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
SaveDimRedux
}
