% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification_Seurat.R
\name{SaveDimRedux_SERIII}
\alias{SaveDimRedux_SERIII}
\title{A Title}
\usage{
SaveDimRedux_SERIII(seuratObj, reductions = c("pca", "tsne", "umap"),
  save.path = NA, maxPCAcomps = 10, nameID = "")
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
