% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{SetIfNull}
\alias{SetIfNull}
\title{A Title}
\usage{
SetIfNull(x, default)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
