% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{SingleRmySerObj}
\alias{SingleRmySerObj}
\title{SingleR my Seurat Object}
\usage{
SingleRmySerObj(seuratObj = NULL, dataset = "hpca", assay = NULL,
  singleRSavePath = NULL)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{dataset}{The dataset (see singleR docs) to use as a reference}

\item{assay}{The assay in the seurat object to use}

\item{singleRSavePath}{If provided, the singleR object will be saved here}
}
\value{
The modified Seurat object
}
\description{
Compute SingleR classification on a Seurat object from path or direcly
}
\keyword{Classification}
\keyword{SingleR}
