% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{Tabulate_SingleRClassLabs}
\alias{Tabulate_SingleRClassLabs}
\title{Tabulate SingleR Class Lables}
\usage{
Tabulate_SingleRClassLabs(seuratObject, plotIndividually = F)
}
\arguments{
\item{seuratObject}{a Seurat object, but if path given, path is prioritized.}

\item{plotIndividually}{If true, two separate plots will be printed.  Otherwise a single plot wil be printed with one above the other}
}
\description{
Tabulate Seurobject with SingleR class lables
}
\keyword{Classification}
\keyword{SingleR}
\keyword{Tabulate}
