% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{transposedt}
\alias{transposedt}
\title{A Title}
\usage{
transposedt(dt, varlabel = "myVar")
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
