% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{UniformSampleDF_FacPor}
\alias{UniformSampleDF_FacPor}
\title{UniformSampleDF_FacPor}
\usage{
UniformSampleDF_FacPor(x, ClassF, p)
}
\value{
index of the uniformly sampled samples
}
\description{
uniformly sample a dataframe based on factor and porportion
}
\keyword{sample,}
\keyword{uniform}
