% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{WilcoxDETest}
\alias{WilcoxDETest}
\title{A Title}
\usage{
WilcoxDETest(object, cells.1, cells.2, verbose = TRUE, ...)
}
\arguments{
\item{object, }{A Seurat object.}

\item{cells.1}{group 1}

\item{cells.2}{group 2}

\item{verbose}{Set true for a more verbose output}
}
\value{
A modified Seurat object.
}
\description{
A description
}
