% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III_functions.R
\name{writeCellBarcodes}
\alias{writeCellBarcodes}
\title{FunctionName.}
\usage{
writeCellBarcodes(seuratObj, file)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
