% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{writeCellBarcodes}
\alias{writeCellBarcodes}
\title{A Title}
\usage{
writeCellBarcodes(seuratObj, file)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
