% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{WriteCellBarcodes}
\alias{WriteCellBarcodes}
\title{WriteCellBarcodes}
\usage{
WriteCellBarcodes(seuratObj, file)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{seurObj, }{A Seurat object.}

\item{The}{file to which barcodes will be written}
}
\value{
A modified Seurat object.
}
\description{
Writes a table of cell barcodes to the provided file
}
