% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{WriteSummaryMetrics}
\alias{WriteSummaryMetrics}
\title{WriteSummaryMetrics}
\usage{
WriteSummaryMetrics(seuratObj, file)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\description{
WriteSummaryMetrics
}
