% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{writeSummaryMetrics}
\alias{writeSummaryMetrics}
\title{FunctionName.}
\usage{
writeSummaryMetrics(seuratObj, file)
}
\arguments{
\item{SeurObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
FunctionName.
}
\examples{
MyFxName_SER(SeurObj=SO)
}
